// tools/workspace-plugin/src/generators/react-library/schema.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ReactLibraryGeneratorSchema {
  /**
   * Library name (without @fluentui scope prefix)
   */
  name: string;
  /**
   * Team owning the library. Will be written in CODEOWNERS
   */
  owner: string;
  /**
   * v9 library kind either embracing converged patterns(standard) or using griffel only with old framework patterns(compat)
   */
  kind?: 'standard' | 'compat';
}